% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/above_quantile.R
\name{above_quantile}
\alias{above_quantile}
\title{Sign of the sample quantile}
\usage{
above_quantile(x, alpha = 0.05)
}
\arguments{
\item{x}{numerical vector}

\item{alpha}{significance level}
}
\value{
The function returns \code{TRUE} if zero is greater than the
(1-\code{alpha})-quantile of \code{x}.
}
\description{
Internal function. It checks whether zero is greater than the sample quaantile
of a vector.
}
\examples{
Y <- rnorm(100, mean=-2, sd=1)
hist(Y)
above_quantile(Y, 0.05)

Y <- rnorm(100, mean=0, sd=1)
hist(Y)
above_quantile(Y, 0.05)
}
\author{

}
