% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_hp.R
\name{check_hp}
\alias{check_hp}
\title{Condition check: subset under test}
\usage{
check_hp(x, N)
}
\arguments{
\item{x}{an R object}

\item{N}{maximum index}
}
\value{
The function returns \code{TRUE} if \code{x} is a vector, and all its elements
are integers greater or equal to 1 and smaller or equal to \code{N}.
}
\description{
Internal function. It checks whether an object is a well-defined
vector of indices to be tested.
}
\examples{
S <- c(5, 7, 1, 9)
check_hp(S, 20)
check_hp(S, 8)

S <- c(1.5, 7, 1, 9)
check_hp(S, 20)
}
\author{

}
