% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_indices.R
\name{check_indices}
\alias{check_indices}
\title{Condition check: matrix of indices}
\usage{
check_indices(x, R, C)
}
\arguments{
\item{x}{matrix of finite numbers}

\item{R}{number of rows}

\item{C}{number of columns}
}
\value{
The function returns \code{TRUE} if \code{x} has \code{R} rows
and \code{C} columns, and each row contains an unordered sequence of
integers from 1 to \code{C}.
}
\description{
Internal function. It checks whether a matrix is a well-defined
matrix of indices with given dimensions.
}
\examples{
M <- matrix(c(1, 3, 2, 2, 3, 1, 2, 1, 3, 2, 3, 1), ncol=3, byrow=TRUE)
check_indices(M, R=4, C=3)
M <- matrix(c(9, 3, 2, 2, 3, 1, 2, 1, 3, 2, 3, 1), ncol=3, byrow=TRUE)
check_indices(M, R=4, C=3)
}
\author{

}
