% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_matrix.R
\name{check_matrix}
\alias{check_matrix}
\title{Condition check: matrix of finite numbers}
\usage{
check_matrix(x)
}
\arguments{
\item{x}{an R object}
}
\value{
The function returns \code{TRUE} if \code{x} is a matrix and all its elements
are finite numbers.
}
\description{
Internal function. It checks whether an object is a matrix of finite numbers.
}
\examples{
M <- matrix(c(0, 1, 2, 3), nrow=2)
check_matrix(M)
}
\author{

}
