% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incr_check.R
\name{incr_check}
\alias{incr_check}
\title{Increasing check}
\usage{
incr_check(Ds, Dc, w = ncol(Dc) - 1, alpha = 0.05)
}
\arguments{
\item{Ds}{numerical vector, sum of the centered test statistics of the indices under testing}

\item{Dc}{matrix of the centered test statistics of the indices not under testing}

\item{w}{maximum size to be considered,
should be equal to the number of columns of \code{Dc} minus 1.}

\item{alpha}{significance level}
}
\value{
The function returns a list with the following objects:
\code{rej} (\code{FALSE} if a non-rejection has been found) and
\code{indecisive} (vector of sizes with indecisive outcome).
}
\description{
Internal function. It checks the bounds for different superset sizes,
considered in increasing order from 1 up to a given maximum.
}
\examples{
M <- matrix(c(0, 1, 2, 3), nrow=2)
check_matrix(M)
}
\author{

}
